feat: car detail page fixing
