+++ components/pages/license/LicenseWrapper.module.css
@@ -1,4 +1,11 @@
 .licenseWrap {
-  opacity: 0.9;
-  padding: 0 2.5rem;
+  padding: 2rem 2.5rem;
+  position: fixed;
+  top: 50px;
+  left: 0;
+  z-index: 300;
+  background: #fff;
+  border-radius: 1rem 1rem 0 0;
+  width: 100%;
+  height: 100vh;
 }
