--- pages/car/[cid]/license.tsx
@@ -1,12 +1,0 @@
-import LicenseWrapper from "@/components/pages/license/LicenseWrapper";
-
-export default function License() {
- 
-  return (
-    <main>
-       <LicenseWrapper />
-    </main>
-  );
-}
-
-License.auth = true;
